feat(migration): adopt native Postgres enums + sqlx macros, fix scraping issues
Main changes (Cursor migration + manual fixes):
- Replace all string-cast WHERE comparisons (type::text = 'MOVIE',
  status::text = $N,
  format!("m.type::text = '{...}'")) with native enum binds — no cast,
  no injection risk
- Convert db/meta.rs Internal branch to query_as! macro (compile-time
  SQL verification,
  mirrors External branch; collapses duplicated base_sql)
- Port contribution bot handlers (bot/) to Rust with full parity to
  Python
- Update all DB query result types to use MediaId/UserId/MediaType
  newtypes so wrong
  types cause compile errors rather than silent runtime failures
- Add 13 new .sqlx/ entries for offline SQLX compile-time query
  verification
- Fix configure.rs: preserve secret_str through redirect to
  /app/configure?secret_str=
- Fix metrics_middleware.rs: RAII InFlightGuard prevents stuck in-flight
  counter on
  client disconnect / future cancellation
- Fix easynews.rs: 401 short-circuits all remaining queries
  (user-specific cred failure)
- Fix torbox_search.rs: 429 skips title query and returns early
  (per-token rate limit)
- Fix public_indexer_registry.rs: sort non-CF indexers first so byparr
  sites run last
- Disable CF bypass (byparr) in live API fan_out — Chromium sessions
  only in bg workers
- Add source health gate (NZBIndex + Binsearch) — 429s record failures;
  gate blocks
  further calls once success rate falls below threshold
- Fix /metrics column bug (media_type → type); replace COUNT(*) full
  table scans on
  large stream tables with pg_stat_user_tables n_live_tup estimates
- Remove unauthenticated /metrics root route; keep /api/v1/metrics under
  api_key_middleware
- Log Zilean response body on JSON parse failure (was silently dropping
  error context)
- Add Grafana + Prometheus dev stack (docker-compose-grafana-dev.yml,
  prometheus-dev.yml)

+++ backend/src/routes/content/torrent_import.rs
@@ -506,7 +506,7 @@ pub async fn analyze_magnet(
             .and_then(|id| {
                 matches
                     .iter()
-                    .find(|m| m["media_id"].as_i64() == Some(id))
+                    .find(|m| m["media_id"].as_i64() == Some(i64::from(i32::from(id))))
                     .cloned()
             })
         } else {
